feat: move graphapi calls into subdir
