+++ src/Frontend/Views/Portal/PortalMainView.tsx
@@ -3,8 +3,8 @@ import { IconType } from '@darkforest_eth/ui';
 import React, { useEffect, useState } from 'react';
 import { Redirect, Route, Switch, useHistory } from 'react-router-dom';
 import styled from 'styled-components';
-import { loadAccountData } from '../../../Backend/Network/AccountApi';
-import { loadRecentMaps } from '../../../Backend/Network/MapsApi';
+import { loadAccountData } from '../../../Backend/Network/GraphApi/AccountApi';
+import { loadRecentMaps } from '../../../Backend/Network/GraphApi/MapsApi';
 import { Btn } from '../../Components/Btn';
 import Button from '../../Components/Button';
 import { Dropdown, DropdownItem } from '../../Components/Dropdown';
